Decompose the tab switching for better readability
